refactor(ui): split markdown renderer modules
- split markdown renderer into parser/render/lists/code/metadata modules
- keep public entry points in markdown/mod.rs with minimal glue wiring
- move markdown tests into behavior-focused modules under markdown/tests
- add shared markdown test helpers for line/span assertions
- update README architecture section for modular markdown layout
