+++ src/ui/markdown/tests/mod.rs
@@ -1,0 +1,6 @@
+mod helpers;
+mod lists;
+mod misc;
+mod syntax_spans;
+mod tables;
+mod wrapping;
